% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateSubdata.R
\name{aggregateSubdata}
\alias{aggregateSubdata}
\title{Data Aggregation Function}
\usage{
aggregateSubdata(subdata, meta.data = getMeta1(subdata))
}
\arguments{
\item{subdata}{Subset of data}

\item{meta.data}{meta object created by function which will be passed as the default argument}
}
\description{
This function aggregates a subset of data based on its corresponding meta object
This will be called upon later for funnel plot creation
}
