% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getES.R
\name{getES}
\alias{getES}
\title{Effect Size Computer}
\usage{
getES(subdata)
}
\arguments{
\item{subdata}{The subset of data to be used to obtain these two tables from}
}
\description{
This function computes the effect sizes of a data set, subsetting by type
}
