% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeta.R
\name{getMeta}
\alias{getMeta}
\title{Meta Coefficients Function}
\usage{
getMeta(subdata, task.variables)
}
\arguments{
\item{subdata}{Subset of data}

\item{task.variables}{Vector of strings, default is a function of the subset of data}
}
\description{
This function analyzes a dataset to create a meta object
}
