% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeta1.R
\name{getMeta1}
\alias{getMeta1}
\title{Meta Coefficients Function}
\usage{
getMeta1(subdata, task.variables)
}
\arguments{
\item{subdata}{Subset of data}

\item{task.variables}{Vector of strings, default is a function of the subset of data}
}
\description{
This function analyzes a dataset to create a meta object
}
