% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRESandES.R
\name{getRESandES}
\alias{getRESandES}
\title{RES and ES Computer}
\usage{
getRESandES(subdata)
}
\arguments{
\item{subdata}{The subset of data to be used to obtain these two tables from}
}
\description{
This function computes the res and es of a data set, subsetting by type
}
