% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubset.R
\name{getSubset}
\alias{getSubset}
\title{Row extracting function}
\usage{
getSubset(type.v, data = data)
}
\arguments{
\item{data}{The dataset to be used, default name is data}

\item{type}{The type of data to be subsetted}
}
\description{
This function extracts rows corresponding to a String variable and creates a subset
}
\keyword{subset}
