% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaskVariables.R
\name{getTaskVariables}
\alias{getTaskVariables}
\title{Get Task Variables}
\usage{
getTaskVariables(subdata)
}
\arguments{
\item{subdata}{The subset of data have task variables extracted from}
}
\description{
This function extracts the task and variable columns from a dataset, and combines the two into a new vector
}
