% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaregTCB.R
\name{metaregTCB}
\alias{metaregTCB}
\title{Meta Regression Function}
\usage{
metaregTCB(subdata, meta.data)
}
\arguments{
\item{subdata}{Subset of data}

\item{meta.data}{Meta object of the subdata}
}
\description{
Computes unmoderated meta regression of a subset
}
