% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunc.r
\name{plot.funn}
\alias{plot.funn}
\title{Funnel Plot Generator}
\usage{
\method{plot}{funn}(subdata, meta.data)
}
\arguments{
\item{subdata}{Subset of data}

\item{meta.data}{Meta object corresponding to the source}
}
\description{
This function creates an interactive funnel plot with a meta object
}
