% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunn.r
\name{plotfunn}
\alias{plotfunn}
\title{Funnel Plot Generator}
\usage{
plotfunn(meta.data)
}
\arguments{
\item{meta.data}{Meta object corresponding to the source}
}
\description{
This function creates an interactive funnel plot with a meta object
}
