% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaregTCB.R
\name{metaregTCB}
\alias{metaregTCB}
\title{Meta Regression Function}
\usage{
metaregTCB(meta.data)
}
\arguments{
\item{meta.data}{Meta object of the subdata}

\item{subdata}{Subset of data}
}
\description{
Computes unmoderated meta regression of a subset
}
